pyproject.toml
[project]
name = "dependency-scanner-tool-3"
version = "0.1.3"
description = "A tool for scanning and classifying project dependencies"
readme = "README.md"
authors = [
    {name = "Project Team", email = "team@example.com"}
]
requires-python = ">=3.8"
dependencies = [
    "tomli>=2.0.1",  # For TOML parsing
    "packaging>=23.0",  # For Python package version parsing
    "pyyaml>=6.0",  # For YAML configuration files
    "click>=8.0.0",  # For CLI interface
    "jinja2>=3.0.0",
    "fastapi>=0.104.0",  # For REST API
    "uvicorn[standard]>=0.24.0",  # For ASGI server
    "python-multipart>=0.0.6",  # For form data handling
    "gitpython>=3.1.40",  # For secure Git operations
    "python-jose[cryptography]>=3.3.0",  # For JWT/auth support
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "httpx>=0.25.0",  # For testing FastAPI
    "pytest-asyncio>=0.21.0",  # For async test support
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true

[project.scripts]
dependency-scanner-tool-3 = "dependency_scanner_tool.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/dependency_scanner_tool"]